requirements.txt
langchain
langchain-community
openai
psycopg2-binary
pgvector
python-dotenv
pymupdf
tiktoken
tabulate